+++ src/main/java/com/braintreegateway/DisputeTransaction.java
@@ -10,6 +10,7 @@ public final class DisputeTransaction {
     private final BigDecimal amount;
     private final Calendar createdAt;
     private final String id;
+    private final String installmentCount;
     private final String orderId;
     private final String paymentInstrumentSubtype;
     private final String purchaseOrderNumber;
@@ -18,6 +19,7 @@ public final class DisputeTransaction {
         amount = node.findBigDecimal("amount");
         createdAt = node.findDateTime("created-at");
         id = node.findString("id");
+        installmentCount = node.findString("installment-count");
         orderId = node.findString("order-id");
         paymentInstrumentSubtype = node.findString("payment-instrument-subtype");
         purchaseOrderNumber = node.findString("purchase-order-number");
@@ -33,6 +35,10 @@ public final class DisputeTransaction {
 
     public String getId() {
         return id;
+    }
+
+    public String getInstallmentCount() {
+        return installmentCount;
     }
 
     public String getOrderId() {